Make the strict clippy gate pass, and fix what it found
Two of these were real. Paced subscription replay took a ReplayPacing and
never read it, so a reconnect re-sent every subscription at once — the
pacing loop landed in the signature but not the body. And a doc comment
split during an extraction left half of it describing the type below it
rather than the function it was written for.

The rest is lint debt: inlined format arguments, redundant f64 casts left
by the fractional-position change, and a file-level allow for
too_many_arguments. A function that encodes a wire message takes the
fields the protocol says that message carries; splitting one to satisfy an
arity count would hide the correspondence the message definition depends on.

+++ src/api/client/account.rs
@@ -54,7 +54,7 @@ impl EClient {
                     ..Default::default()
                 });
             let avg_cost = pi.avg_cost as f64 / PRICE_SCALE_F;
-            wrapper.position(&self.account_id, &c, pi.position as f64, avg_cost);
+            wrapper.position(&self.account_id, &c, pi.position, avg_cost);
         }
         wrapper.position_end();
     }
